2.0版功能已完成， 开始测试

+++ src/main/java/org/yi/spider/helper/FileHelper.java
@@ -30,7 +30,7 @@ public class FileHelper {
      */
 	public static void writeTxtFile(NovelEntity article, ChapterEntity chapter, String content) {
 		
-		int novelNo = article.getNovelNo();
+		int novelNo = article.getNovelNo().intValue();
 		int subDir = novelNo/1000;
         String localPath = GlobalConfig.localSite.getTxtDir() + FileUtils.FILE_SEPARATOR
 		        		+ subDir + FileUtils.FILE_SEPARATOR
